TradeBotsAI/requirements.txt
optuna>=3.6
pillow>=10.0
pytesseract>=0.3.10
mss>=9.0
keyboard>=0.13.5
pyautogui>=0.9.54
